add custom protocol example & clean up app.c
